+++ bin/c.js
@@ -12,6 +12,7 @@ program
     .command('gc <command>', 'Shortcuts to help with gcloud.')
     .command('hosts <command>', 'Shortcuts to help with hosts management.')
     .command('kc <command>', 'Shortcuts to help with kubectrl.')
+    .command('fs <command>', 'Shortcuts to help with the file system.')
     .command('mk <command>', 'Shortcuts to control Minikube.')
     .command('mongo <command>', 'Shortcuts to help with the database.')
     .command('npm <command>', 'Shortcuts to help with NPM.')